--- airbyte-webapp/src/components/Spinner/index.tsx
@@ -1,4 +1,0 @@
-import Spinner from "./Spinner";
-
-export default Spinner;
-export { Spinner };
