+++ airbyte-webapp/src/components/index.tsx
@@ -15,7 +15,6 @@ export * from "./MainPageWithScroll";
 export * from "./Modal";
 export * from "./PageTitle";
 export * from "./SimpleTableComponents";
-export * from "./Spinner";
 export * from "./StatusIcon";
 export * from "./TextWithHTML";
 export * from "./ProgressBar";
