+++ airbyte-webapp/src/components/ui/Spinner/Spinner.module.scss
@@ -1,4 +1,4 @@
-@use "../../scss/colors";
+@use "../../../scss/colors";
 
 @keyframes spin {
   from {
