+++ airbyte-webapp/src/components/ui/Spinner/Spinner.test.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 
-import Spinner from "./Spinner";
+import { Spinner } from "./Spinner";
 
 describe("<Spinner />", () => {
   it("should render without crash", () => {
