+++ airbyte-webapp/src/components/ui/Spinner/Spinner.tsx
@@ -7,8 +7,6 @@ interface SpinnerProps {
   small?: boolean;
 }
 
-const Spinner: React.FC<SpinnerProps> = ({ small }) => (
+export const Spinner: React.FC<SpinnerProps> = ({ small }) => (
   <div className={classnames(styles.spinner, { [styles.small]: small })} />
 );
-
-export default Spinner;
