+++ airbyte-webapp/src/views/Connection/ConnectionForm/components/CreateControls.tsx
@@ -2,7 +2,8 @@ import React from "react";
 import { FormattedMessage } from "react-intl";
 import styled from "styled-components";
 
-import { Button, Spinner, StatusIcon } from "components";
+import { Button, StatusIcon } from "components";
+import { Spinner } from "components/ui/Spinner";
 
 interface CreateControlsProps {
   isSubmitting: boolean;
